Complete native isolation profile and test integration
