+++ src-tauri/gen/schemas/capabilities.json
@@ -1,1 +1,1 @@
-{"default":{"identifier":"default","description":"Default capabilities for StrokeMotion","local":true,"windows":["main"],"permissions":["core:default","core:window:allow-start-dragging","core:window:allow-close","core:window:allow-destroy","shell:allow-open","updater:default","process:allow-restart",{"identifier":"shell:allow-execute","allow":[{"args":true,"name":"binaries/ffmpeg","sidecar":true}]},"fs:default","fs:allow-write-file","fs:allow-write-text-file","fs:allow-read-text-file","fs:allow-read-file","fs:allow-mkdir","fs:allow-read-dir","fs:allow-remove","fs:allow-exists","fs:allow-rename","fs:scope-app-recursive",{"identifier":"fs:scope","allow":[{"path":"$APPCACHE/**"},{"path":"$TEMP/**"},{"path":"$HOME/**"},{"path":"$DOCUMENT/**"},{"path":"$DESKTOP/**"},{"path":"$DOWNLOAD/**"}]},"dialog:default","dialog:allow-save","dialog:allow-open",{"identifier":"http:default","allow":[{"url":"http://*/*"},{"url":"https://*/*"}]}]}}
+{"default":{"identifier":"default","description":"Default capabilities for StrokeMotion","local":true,"windows":["main"],"permissions":["core:default","core:window:allow-start-dragging","core:window:allow-close","core:window:allow-destroy","shell:allow-open","updater:default","process:allow-restart",{"identifier":"shell:allow-execute","allow":[{"name":"binaries/ffmpeg","sidecar":true,"args":true}]},"fs:default","fs:allow-write-file","fs:allow-write-text-file","fs:allow-read-text-file","fs:allow-read-file","fs:allow-mkdir","fs:allow-read-dir","fs:allow-remove","fs:allow-exists","fs:allow-rename","fs:scope-app-recursive",{"identifier":"fs:scope","allow":[{"path":"$APPCACHE/**"},{"path":"$TEMP/**"},{"path":"$HOME/**"},{"path":"$DOCUMENT/**"},{"path":"$DESKTOP/**"},{"path":"$DOWNLOAD/**"}]},"dialog:default","dialog:allow-save","dialog:allow-open","dialog:allow-ask","dialog:allow-confirm",{"identifier":"http:default","allow":[{"url":"http://*/*"},{"url":"https://*/*"}]}]}}
