+++ tests/nemo-desktop-harness.test.cjs
@@ -1,0 +1,4 @@
+'use strict';
+
+// Exercise harness failure controls in ordinary Node checks, without launching a desktop.
+require('./desktop/unit/native-harness.test.cjs');
